make cache quick switch default
